requirements.txt
torch
torchvision
transformers
faiss-cpu
opencv-python
Pillow
numpy
tqdm
qdrant-client
albumentations
streamlit
pandas
